Notes App

+++ to_do/backend/src/controllers/notesControllers.ts
@@ -2,10 +2,15 @@ import { RequestHandler } from "express"
 import NotesModel from "../models/node"
 import createHttpError from "http-errors"
 import mongoose from "mongoose"
+import {assertIsDefined} from "../util/assertIsDefined"
 
 export const getNotes: RequestHandler = async (req, res, next) => {
+    const authenticatedUserId = req.session.userId
+
     try {
-        const notes = await NotesModel.find().exec()
+        assertIsDefined(authenticatedUserId)
+
+        const notes = await NotesModel.find({userId: authenticatedUserId}).exec()
         res.status(200).json(notes)
     } catch (error) {
         next(error)
@@ -14,7 +19,11 @@ export const getNotes: RequestHandler = async (req, res, next) => {
 
 export const getNote: RequestHandler = async (req, res, next) => {
     const noteId = req.params.noteId
+    const authenticatedUserId = req.session.userId
+
     try {
+        assertIsDefined(authenticatedUserId)
+
         if (!mongoose.isValidObjectId(noteId)) {
             throw createHttpError(400, "Invalid note ID")
         }
@@ -24,6 +33,11 @@ export const getNote: RequestHandler = async (req, res, next) => {
         if (!note) {
             throw createHttpError(404, "Note not found")
         }
+
+        if(!note.userId.equals(authenticatedUserId)){
+            throw createHttpError(401, "unAuthorized access")
+        }
+
         res.status(200).json(note)
     } catch (error) {
         next(error)
@@ -38,14 +52,16 @@ interface CreateNoteBody {
 export const createNotes: RequestHandler<unknown, unknown, CreateNoteBody, unknown> = async (req, res, next) => {
     const title = req.body.title
     const text = req.body.text
+    const authenticatedUserId = req.session.userId
 
     try {
-
+        assertIsDefined(authenticatedUserId)
         if (!title) {
             throw createHttpError(400, "Note must have a title")
         }
 
         const newNotes = await NotesModel.create({
+            userId: authenticatedUserId,
             title: title,
             text: text,
         })
@@ -69,8 +85,11 @@ export const updateNotes: RequestHandler<UpdateNoteId, unknown, UpdateNoteBody, 
     const noteId = req.params.noteId
     const newTitle = req.body.title
     const newText = req.body.text
+    const authenticatedUserId = req.session.userId
 
     try {
+        assertIsDefined(authenticatedUserId)
+
         if(!mongoose.isValidObjectId(noteId)){
             throw createHttpError(400, "Invalid Note ID")
         }
@@ -83,6 +102,9 @@ export const updateNotes: RequestHandler<UpdateNoteId, unknown, UpdateNoteBody, 
 
         if(!note){
             throw createHttpError(404, "Note not found")
+        }
+        if(!note.userId.equals(authenticatedUserId)){
+            throw createHttpError(401, "unAuthorized access")
         }
 
         note.title = newTitle
@@ -98,8 +120,10 @@ export const updateNotes: RequestHandler<UpdateNoteId, unknown, UpdateNoteBody, 
 
 export const deleteNote: RequestHandler = async (req, res, next) => {
     const noteId = req.params.noteId
+    const authenticatedUserId = req.session.userId
 
     try {
+        assertIsDefined(authenticatedUserId)
         if(!mongoose.isValidObjectId(noteId)){
             throw createHttpError(400, "Invalid Note ID")
         }
@@ -108,6 +132,9 @@ export const deleteNote: RequestHandler = async (req, res, next) => {
 
         if(!note){
             throw createHttpError(404, "Note not found")
+        }
+        if(!note.userId.equals(authenticatedUserId)){
+            throw createHttpError(401, "unAuthorized access")
         }
 
         await note.deleteOne()
